backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
celery>=5.3.6
redis>=5.0.3
sqlalchemy>=2.0.29
asyncpg>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
# AI & LangChain
google-generativeai>=0.5.0
langchain>=0.1.16
# Image Processing for Security/Masking
opencv-python-headless>=4.9.0.80
numpy>=1.26.4
# Testing integration
pytest>=7.4.4
playwright>=1.41.0
